+++ taxsystem/api/filters.py
@@ -30,6 +30,7 @@ from taxsystem.models.corporation import (
     CorporationOwner,
 )
 from taxsystem.models.helpers.textchoices import (
+    ActionType,
     AdminActions,
 )
 from taxsystem.providers import AppLogger
@@ -162,6 +163,7 @@ class FilterApiEndpoints:
             owner.admin_log_model(
                 user=request.user,
                 owner=owner,
+                target=ActionType.FILTER,
                 action=AdminActions.DELETE,
                 comment=msg,
             ).save()
@@ -278,6 +280,7 @@ class FilterApiEndpoints:
             owner.admin_log_model(
                 user=request.user,
                 owner=owner,
+                target=ActionType.FILTER_SET,
                 action=AdminActions.DELETE,
                 comment=msg,
             ).save()
@@ -333,6 +336,14 @@ class FilterApiEndpoints:
                 filter_set=filter_set,
                 enabled=filter_set.enabled,
             )
+            # Log the change in Admin History
+            owner.admin_log_model(
+                user=request.user,
+                owner=owner,
+                target=ActionType.FILTER_SET,
+                action=AdminActions.CHANGE,
+                comment=msg,
+            ).save()
 
             # Return success response
             return 200, {"success": True, "message": msg}